Added utility function for renaming collapsed blocks and collapsed parents.
