+++ appinventor/blocklyeditor/src/blocks/lexical-variables.js
@@ -137,12 +137,7 @@ Blockly.Blocks['lexical_variable_get'] = {
     // console.log("Renaming lexical variable from " + oldName + " to " + newName);
     if (oldName === this.getFieldValue('VAR')) {
         this.setFieldValue(newName, 'VAR');
-        if (this.isCollapsed()) {
-        var COLLAPSED_INPUT_NAME = '_TEMP_COLLAPSED_INPUT';
-        this.removeInput(COLLAPSED_INPUT_NAME);
-        var text = this.toString(Blockly.COLLAPSE_CHARS);
-        this.appendDummyInput(COLLAPSED_INPUT_NAME).appendField(text);
-      }
+        Blockly.Blocks.Utilities.renameCollapsed(this, 0);
     }
   },
   renameFree: function (freeSubstitution) {
