+++ appinventor/blocklyeditor/src/blocks/procedures.js
@@ -252,14 +252,11 @@ Blockly.Blocks['procedures_defnoreturn'] = {
 
       var callers = Blockly.Procedures.getCallers(procName, procWorkspace);
       for (var x = 0; x < callers.length; x++) {
-          var block = callers[x];
-          if (block.isCollapsed()) {
-          var COLLAPSED_INPUT_NAME = '_TEMP_COLLAPSED_INPUT';
-          block.removeInput(COLLAPSED_INPUT_NAME);
-          var text = block.toString(Blockly.COLLAPSE_CHARS);
-          block.appendDummyInput(COLLAPSED_INPUT_NAME).appendField(text);
-          block.moveInputBefore(COLLAPSED_INPUT_NAME, 'ARG0');
-        }
+           var block = callers[x];
+           Blockly.Blocks.Utilities.renameCollapsed(block, 0);
+      //     if (block.isCollapsed()) {
+      //     block.moveInputBefore(COLLAPSED_INPUT_NAME, 'ARG0');
+      //   }
       }
 
       // 2. If there's an open mutator, change the name in the corresponding slot.
@@ -763,12 +760,7 @@ Blockly.Blocks['procedures_callnoreturn'] = {
   renameProcedure: function(oldName, newName) {
     if (Blockly.Names.equals(oldName, this.getFieldValue('PROCNAME'))) {
       this.setFieldValue(newName, 'PROCNAME');
-      if (this.isCollapsed()) {
-        var COLLAPSED_INPUT_NAME = '_TEMP_COLLAPSED_INPUT';
-        this.removeInput(COLLAPSED_INPUT_NAME);
-        var text = this.toString(Blockly.COLLAPSE_CHARS);
-        this.appendDummyInput(COLLAPSED_INPUT_NAME).appendField(text);
-      }
+      Blockly.Blocks.Utilities.renameCollapsed(this, 0);
     }
   },
   // [lyn, 10/27/13] Renamed "fromChange" parameter to "startTracking", because it should be true in any situation
